Continue work from gruenfink device

+++ trace_for_guess/add_precc_precl.py
@@ -15,18 +15,24 @@ def add_precc_and_precl_to_prect(precc_file, precl_file, prect_file):
     Args:
         precc_file: The TraCE-21ka NetCDF file with the PRECC variable.
         precl_file: The TraCE-21ka NetCDF file with the PRECL variable.
-        prect_file: The output file.
+        prect_file: The output file (will *not* be overwritten).
+
+    Returns:
+        The PRECT file (=`prect_file`).
 
     Raises:
         FileNotFoundError: Either the PRECC or the PRECL file couldn’t be
             found.
         RuntimeError: The `ncrename` command has failed or produced not output.
     """
-    cprint("Adding PRECC and PRECL to PRECT: '%s'" % prect_file, "green")
+    cprint("Adding PRECC and PRECL to PRECT: '%s'" % prect_file, 'yellow')
     if not os.path.isfile(precc_file):
         raise FileNotFoundError("Could not find PRECC file: '%s'" % precc_file)
     if not os.path.isfile(prect_file):
         raise FileNotFoundError("Could not find PRECT file: '%s'" % prect_file)
+    if os.path.isfile(prect_file):
+        cprint('PRECT file already exists. Skipping.', 'cyan')
+        return prect_file
     # First copy PRECC file into output location so that we can rename its
     # variable to match the variable of the other operand.
     shutil.copy2(precc_file, prect_file)
@@ -37,10 +43,8 @@ def add_precc_and_precl_to_prect(precc_file, precl_file, prect_file):
     if status != 0:
         raise RuntimeError('Command `ncrename` failed.')
     # Now we can add the matching variable name "PRECL".
-    status = run('ncbo', [
-        '--overwrite', '--op_typ=add', f'-o {prect_file}', precl_file,
-        prect_file
-    ])
+    status = run('ncbo', ['--op_typ=add', f'-o {prect_file}', precl_file,
+                          prect_file])
     if status != 0:
         if os.path.isfile(prect_file):
             os.remove(prect_file)
@@ -61,3 +65,5 @@ def add_precc_and_precl_to_prect(precc_file, precl_file, prect_file):
             os.remove(prect_file)
         raise RuntimeError(f"Failed to set attribute for file '{prect_file}"
                            "with `ncrename`.")
+    cprint(f"Successfully created '{out_file}'.", 'green')
+    return prect_file
